apple-codesign: convert extract command to use sub-commands
This allows us to capture documentation for each data type next to
its definition instead of in separate help text.

Since we can't have a global argument following the sub-command, we
had to invent a custom data structure for holding it and had to
override the usage text to preserve the invocation form.

While we were here we also moved the command dispatch into the
struct, as is the modern convention.